requirements.txt
# # # ============================================================
# # # MathAI Backend Dependencies
# # # ============================================================

# # # Web Framework
# # fastapi==0.104.1
# # uvicorn[standard]==0.24.0
# # python-multipart==0.0.6
# # websockets==12.0

# # # CORS and middleware
# # python-dotenv==1.0.0
# # pydantic==2.5.0
# # pydantic-settings==2.1.0

# # # AI & ML
# # google-generativeai==0.3.2
# # sentence-transformers==2.2.2
# # torch==2.9.0
# # transformers==4.35.2

# # # Vector Database
# # qdrant-client==1.7.0

# # # Web Search
# # tavily-python==0.3.3
# # requests==2.31.0
# # beautifulsoup4==4.12.2

# # # Database
# # sqlalchemy==2.0.23
# # aiosqlite==0.19.0

# # # Data Processing
# # pandas==2.1.3
# # numpy==1.26.2

# # # Utilities
# # python-dateutil==2.8.2
# # pytz==2023.3

# # # Testing
# # pytest==7.4.3
# # pytest-asyncio==0.21.1
# # httpx==0.25.2

# # # Code Quality
# # black==23.11.0
# # isort==5.12.0
# # flake8==6.1.0

# # # Monitoring
# # python-json-logger==2.0.7

# # # Security
# # cryptography==41.0.7

# # ============================================================
# # MathAI Backend Dependencies (Python 3.13 Compatible)
# # ============================================================

# # Web Framework
# fastapi==0.115.5
# uvicorn[standard]==0.34.0
# python-multipart==0.0.20
# websockets==14.1

# # CORS and middleware
# python-dotenv==1.0.1
# pydantic==2.10.5
# pydantic-settings==2.7.1

# # AI & ML (Compatible with Python 3.13)
# google-generativeai==0.8.3
# sentence-transformers==3.3.1
# torch>=2.5.0
# transformers==4.47.1

# # Vector Database
# qdrant-client==1.12.1

# # Web Search
# tavily-python==0.5.0
# requests==2.32.3
# beautifulsoup4==4.12.3

# # Database
# sqlalchemy==2.0.36

# # Data Processing
# pandas==2.2.3
# numpy==2.2.1

# # Testing (optional)
# pytest==8.3.4
# httpx==0.28.1




# ============================================================
# MathAI Backend Dependencies (Python 3.10+ Compatible)
# ============================================================

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# CORS and middleware
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# AI & ML
google-generativeai==0.8.3
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# Vector Database
qdrant-client==1.12.1

# Web Search
tavily-python==0.5.0
requests==2.32.3
beautifulsoup4==4.12.3

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Utilities
python-dateutil==2.9.0

# Testing (optional)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Code Quality (optional)
black==24.10.0
isort==5.13.2
flake8==7.1.1